Added sbt project definition and removed Makefile.
From now (assuming sbt is installed) one can use sbt to build
the project. The best way to work with the project is to enter
sbt's shell just by typing 'sbt' and then using continous mode
for running the app.

Just type:
~run

and sbt will monitor for any source changes, recompile it
and rerun the app.