Added random seed when hashing is enabled.
In previous implementation it was easy to reverse the hash
function because both hash function was known and domain
was known and small (e.g. for person names).

By adding random seed it's significally harder to
reverse hashing function, although still possible.

Random seed is being saved in a file that should
be kept private.