Fixed calculation of correlations.
Calculation of correlations was completely broken because
we were trying to calculate correlations between sequences
of different size.

Different size comes from the fact that answering to all
questions is not required so we have different number of
answers for given question. Thus, we need to calculate
lists of anwers separately for every pair of questions.

It might happen that for given pair of questions list of
answers is empty. It made sense to introduce data structure
that reflects this fact and it's called PartialMatrix.

This way we can separate generic code for displaying any
PartialMatrix from the code that displays given cell.