Improved reading csv files by using javacsv library.
Made use of javacsv library that safely parsers any
csv file. Previous method was really hacky and fragile
so it made sense to replace it.

Moreover, added reading of headers and changed
DataImporter so it uses header names to parse records
and not positions. This makes code slightly easier
to follow and less error-prone.

While implementing these changes DataImporter has
been heavily refactored so the whole code is
a little bit more clear.