Added reading of question stats and included sample data.
Introduced QuestionStats class which holds simple stats
of given question. Those stats include number of
people that are allowed to answer given question and
number of people that have answered given question.

In addition to that added sample data so people not
having an access to real data can easier test this app.

Because of switching to sample data I had to disable
filtering on stats. One might want to tweak sample data
(by making more entries) so filtering could be turned
on again.