CompleteStats holds reference to identifier for given data.
Introduced CompleteStats.of property that holds reference
to identifier (acting as a key) for given piece of data
that Stats are being computed for.

This way we can incorporate key into CompleteStats and vastly
simplify most of types returned by StatsGenerator. It's worth
noting that although we were returning various Maps they were
never used as such and immediatelly projected to Lists of 
pairs. This way it was very easy to get rid of Maps and
return Lists of CompleteStats instead.

This change allowed to simplify code displaying collections of
CompleteStats as they no longer hold ugly pairs.