Various refactorings towards reduction of String usage.
Stats are holding reference to an object identifying
data that Stats are holding. E.g. if Stats are
calculated for given question then 
Stats.of == q where q is that question.

This little change has rather profound impact on the
rest of the code but it leads to reduced usage of String
for identifying things.

Moreover, introduced QuestionInstance class which holds
data related to use of given Question in a given Survey.
Question class itself does not hold reference to id as
meaning of that id is vague at the moment.

Introduced Show[T] type class which defines how to show
given T both in HTML and String form. Provided default
implementation of Show for String and Question classes.
This should be a general approach to displaying data.