Added a few new plots.
Added plot for distribution of number
question answered. Another (stacked) plot
shows how many answers come with comments.

Also, added a plot that shows number of
answers to each question.

Plotting code is in a little mess and should
be refactored in a future to become a bit
more generic.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>