Print info about correlation in some cases.
In a case that sample size for calculating
correlation is small enough (< 100 right now)
we'll print those questions and subset of
subjects that contributed to that correlation.
At the moment this subset consists of 10
elements.

This is useful to track down fishy cases
where calculating correlation migh not make
too much sense.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>